Correção Implementação do JWT e finalmentes da tarefa
