--- authentication/api/jwt.py
@@ -1,29 +1,0 @@
-from rest_framework.authentication import BaseAuthentication
-from rest_framework.exceptions import AuthenticationFailed
-from django.contrib.auth import get_user_model
-import jwt
-from django.conf import settings
-
-User = get_user_model()
-
-class JWTAuthentication(BaseAuthentication):
-    def authenticate(self, request):
-        token = request.headers.get('Authorization')
-
-        if not token:
-            return None
-
-        try:
-            payload = jwt.decode(token, settings.SECRET_KEY, algorithms=['HS256'])
-        except jwt.ExpiredSignatureError:
-            raise AuthenticationFailed('Token expirado')
-        except jwt.InvalidTokenError:
-            raise AuthenticationFailed('Token inválido')
-
-        try:
-            # Aqui você busca o usuário pelo UUID
-            user = User.objects.get(id=payload['user_id'])
-        except User.DoesNotExist:
-            raise AuthenticationFailed('Usuário não encontrado')
-
-        return (user, None)
